+++ docs/CHANGES.md
@@ -6,7 +6,8 @@
 - Made requesting contacts more visible, by placing them at the top of the roster. [jcbrand]
 - `insertIntoPage` method of `ChatBoxView` has been renamed to `insertIntoDOM`,
   to make it the same as the method of `ChatRoomView`. [jcbrand]
-
+- Render error messages received from the server (for undelivered chat
+  messages). [jcbrand]
 
 ## 1.0.3 (2016-06-20)